fix: pin django-polymorphic to v4.0.0a
Updated the django-polymorphic package to a specific alpha version (v4.0.0a) directly from its Git repository. This change ensures compatibility with ongoing project requirements and addresses potential issues with the previously used version. It is important to monitor the stability and updates of this alpha version closely in the project's development environment.